project water main - Final/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain-huggingface==0.0.3
python-dotenv==1.0.0
requests==2.31.0
